requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
validators==0.22.0
python-dotenv==1.0.0